split events out of actions
