+++ liminal/EventHandler.ts
@@ -1,4 +1,4 @@
-import type { ReturnedEvent } from "./actions/actions_common.ts"
-import type { LEvent } from "./LEvent.ts"
+import type { LEvent } from "./events/LEvent.ts"
+import type { ReturnedEvent } from "./events/ReturnedEvent.ts"
 
 export type EventHandler<E extends LEvent = LEvent, T = any> = (event: E | ReturnedEvent<T>) => any
