+++ liminal/Spec.ts
@@ -1,4 +1,4 @@
-import type { LEvent } from "./LEvent.ts"
+import type { LEvent } from "./events/LEvent.ts"
 
 export interface Spec {
   Entry: [keyof any, any]
