+++ liminal/actions/Await.ts
@@ -1,28 +1,10 @@
-import { Action, type EventBase } from "../Action.ts"
+import { Action } from "../Action.ts"
 
-function* await_<T>(value: T): Generator<
-  Action<"await", {
-    Entry: never
-    Event: AwaitResolvedEvent<T>
-    Throw: never
-  }>,
-  Awaited<T>
-> {
-  return yield Action<never>()("await", async (scope) => {
-    value = await value
-    scope.event({
-      type: "await_resolved",
-      value,
-    })
-    return {
-      ...scope,
-      nextArg: value,
-    }
-  })
+function* await_<T>(value: T): Generator<Action<"await", never>, Awaited<T>> {
+  return yield Action("await", async (scope) => ({
+    ...scope,
+    nextArg: await value,
+  }))
 }
 Object.defineProperty(await_, "name", { value: "await" })
 export { await_ as await }
-
-export interface AwaitResolvedEvent<V = unknown> extends EventBase<"await_resolved"> {
-  value: V
-}
