+++ liminal/actions/DisableTool.ts
@@ -1,8 +1,8 @@
-import { Action, type EventBase } from "../Action.ts"
+import { Action } from "../Action.ts"
 import type { Tool } from "../Tool.ts"
 
 export function* disableTool(tool: Tool): Generator<Action<"disable_tool", never>, void> {
-  yield Action<never>()("disable_tool", (scope) => {
+  yield Action("disable_tool", (scope) => {
     scope.event({
       type: "tool_disabled",
       tool: tool.key,
@@ -15,8 +15,4 @@ export function* disableTool(tool: Tool): Generator<Action<"disable_tool", never
       nextArg: undefined,
     }
   })
-}
-
-export interface ToolDisabledEvent<K extends keyof any = keyof any> extends EventBase<"tool_disabled"> {
-  tool: K
 }
