+++ liminal/actions/Embed.ts
@@ -1,15 +1,17 @@
-import { Action, type EventBase } from "../Action.ts"
+import { Action } from "../Action.ts"
+import type { EmbeddedEvent } from "../events/EmbeddedEvent.ts"
+import type { EmbeddingRequestedEvent } from "../events/EmbeddingRequestedEvent.ts"
 import { assert } from "../util/assert.ts"
 
 export function* embed(value: string): Generator<
   Action<"embed", {
     Entry: never
-    Event: EmbeddedEvent | EmbeddingRequestedEvent
+    Event: EmbeddingRequestedEvent | EmbeddedEvent
     Throw: never
   }>,
   Array<number>
 > {
-  return yield Action<never>()("embed", async (scope) => {
+  return yield Action("embed", async (scope) => {
     assert(scope.runEmbed)
     scope.event({
       type: "embedding_requested",
@@ -26,13 +28,4 @@ export function* embed(value: string): Generator<
       nextArg: scope.value,
     }
   })
-}
-
-export interface EmbeddingRequestedEvent extends EventBase<"embedding_requested"> {
-  value: string
-}
-
-export interface EmbeddedEvent extends EventBase<"embedded"> {
-  value: string
-  embedding: Array<number>
 }
