+++ liminal/actions/Emit.ts
@@ -1,4 +1,5 @@
-import { Action, type EventBase } from "../Action.ts"
+import { Action } from "../Action.ts"
+import type { EmittedEvent } from "../events/EmittedEvent.ts"
 import type { JSONValue } from "../util/JSONValue.ts"
 
 export function* emit<K extends keyof any, V extends JSONValue>(key: K, value: V): Generator<
@@ -9,7 +10,7 @@ export function* emit<K extends keyof any, V extends JSONValue>(key: K, value: V
   }>,
   undefined
 > {
-  return yield Action<never>()("emit", (scope) => {
+  return yield Action("emit", (scope) => {
     scope.event({
       type: "emitted",
       key,
@@ -20,11 +21,4 @@ export function* emit<K extends keyof any, V extends JSONValue>(key: K, value: V
       nextArg: undefined,
     }
   })
-}
-
-export interface EmittedEvent<K extends keyof any = keyof any, E extends JSONValue = JSONValue>
-  extends EventBase<"emitted">
-{
-  key: K
-  value: E
 }
