+++ liminal/actions/EnableTool.type-test.ts
@@ -1,11 +1,13 @@
 import type { AssertTrue, IsExact } from "conditional-type-checks"
+import type { ChildEvent } from "../events/ChildEvent.ts"
+import type { EmittedEvent } from "../events/EmittedEvent.ts"
+import type { ToolCalledEvent } from "../events/ToolCalledEvent.ts"
+import type { ToolDisabledEvent } from "../events/ToolDisabledEvent.ts"
+import type { ToolEnabledEvent } from "../events/ToolEnabledEvent.ts"
 import * as L from "../L.ts"
 import { ActorAssertions } from "../testing/ActorAssertions.ts"
-import type { ChildEvent } from "./actions_common.ts"
-import type { ToolDisabledEvent } from "./DisableTool.ts"
 import { emit } from "./Emit.ts"
-import type { EmittedEvent } from "./Emit.ts"
-import { enableTool, type ToolCalledEvent, type ToolEnabledEvent } from "./EnableTool.ts"
+import { enableTool } from "./EnableTool.ts"
 
 type P = typeof P["T"]
 const P = L.object({
