+++ liminal/actions/SetEmbeddingModel.ts
@@ -1,6 +1,6 @@
-import type { EventBase } from "../Action.ts"
 import { Action } from "../Action.ts"
 import type { RunEmbed } from "../adapters.ts"
+import type { EmbeddingModelSetEvent } from "../events/EmbeddingModelSetEvent.ts"
 
 export function* setEmbeddingModel<K extends keyof any>(key: K, runEmbed: RunEmbed): Generator<
   Action<"set_embedding_model", {
@@ -10,7 +10,7 @@ export function* setEmbeddingModel<K extends keyof any>(key: K, runEmbed: RunEmb
   }>,
   void
 > {
-  return yield Action<never>()("set_embedding_model", (scope) => {
+  return yield Action("set_embedding_model", (scope) => {
     scope.event({
       type: "embedding_model_set",
       key,
@@ -21,8 +21,4 @@ export function* setEmbeddingModel<K extends keyof any>(key: K, runEmbed: RunEmb
       runEmbed,
     }
   })
-}
-
-export interface EmbeddingModelSetEvent<K extends keyof any = keyof any> extends EventBase<"embedding_model_set"> {
-  key: K
 }
