+++ liminal/events/EmbeddingRequestedEvent.ts
@@ -1,0 +1,5 @@
+import type { EventBase } from "./_EventBase.ts"
+
+export interface EmbeddingRequestedEvent extends EventBase<"embedding_requested"> {
+  value: string
+}
