+++ liminal/events/InferenceRequestedEvent.ts
@@ -1,0 +1,3 @@
+import type { EventBase } from "./_EventBase.ts"
+
+export interface InferenceRequestedEvent extends EventBase<"inference_requested"> {}
