+++ liminal/events/LanguageModelSetEvent.ts
@@ -1,0 +1,5 @@
+import type { EventBase } from "./_EventBase.ts"
+
+export interface LanguageModelSetEvent<K extends keyof any = keyof any> extends EventBase<"language_model_set"> {
+  key: K
+}
