+++ liminal/events/ReturnedEvent.ts
@@ -1,0 +1,5 @@
+import type { EventBase } from "./_EventBase.ts"
+
+export interface ReturnedEvent<T = any> extends EventBase<"returned"> {
+  value: T
+}
