+++ liminal/events/ThrownEvent.ts
@@ -1,0 +1,5 @@
+import type { EventBase } from "./_EventBase.ts"
+
+export interface ThrownEvent extends EventBase<"thrown"> {
+  thrown: unknown
+}
