+++ liminal/events/_EventBase.ts
@@ -1,0 +1,3 @@
+export interface EventBase<K extends string> {
+  type: K
+}
