+++ liminal/patterns/clear.ts
@@ -1,5 +1,6 @@
 import type { Action } from "../Action.ts"
-import { type MessagesSetEvent, setMessages } from "../actions/SetMessages.ts"
+import { setMessages } from "../actions/SetMessages.ts"
+import type { MessagesSetEvent } from "../events/MessagesSetEvent.ts"
 import type { Message } from "../Message.ts"
 
 export function clear(): Generator<
