+++ liminal/patterns/declareXModel.type-test.ts
@@ -1,7 +1,7 @@
-import type { ChildEvent } from "../actions/actions_common.ts"
-import type { EmbeddingModelSetEvent } from "../actions/SetEmbeddingModel.ts"
-import type { LanguageModelSetEvent } from "../actions/SetLanguageModel.ts"
 import type { RunEmbed, RunInfer } from "../adapters.ts"
+import type { ChildEvent } from "../events/ChildEvent.ts"
+import type { EmbeddingModelSetEvent } from "../events/EmbeddingModelSetEvent.ts"
+import type { LanguageModelSetEvent } from "../events/LanguageModelSetEvent.ts"
 import * as L from "../L.ts"
 import { ActorAssertions } from "../testing/ActorAssertions.ts"
 import { declareEmbeddingModel } from "./declareEmbeddingModel.ts"
